refactor: AttendanceStatus 해당 로직 추가
