+++ src/main/java/domain/DailyRecord.java
@@ -1,17 +1,9 @@
 package domain;
 
-import static domain.AttendanceStatus.ABSENCE;
-import static domain.AttendanceStatus.ATTENDANCE;
-import static domain.AttendanceStatus.LATENESS;
-import static util.constant.Value.ABSENCE_MINUTE;
-import static util.constant.Value.DEFAULT_START_HOUR;
-import static util.constant.Value.LATENESS_MINUTE;
-import static util.constant.Value.MONDAY_START_HOUR;
+import static domain.AttendanceStatus.checkStatus;
 
 import java.time.DayOfWeek;
 import java.time.LocalTime;
-import java.util.Map;
-import util.parser.DateTimeParser;
 
 public class DailyRecord {
 
@@ -20,27 +12,7 @@ public class DailyRecord {
 
     public DailyRecord(LocalTime time, DayOfWeek dayOfWeek) {
         this.time = time;
-        this.status = checkStatus(dayOfWeek);
-    }
-
-    private AttendanceStatus checkStatus(DayOfWeek dayOfWeek) {
-        Map<DayOfWeek, LocalTime> lateTimes = Map.of(DayOfWeek.MONDAY,
-            DateTimeParser.parseIntegerToTime(MONDAY_START_HOUR, LATENESS_MINUTE));
-        Map<DayOfWeek, LocalTime> absentTimes = Map.of(DayOfWeek.MONDAY,
-            DateTimeParser.parseIntegerToTime(MONDAY_START_HOUR, ABSENCE_MINUTE));
-
-        LocalTime lateTime = lateTimes.getOrDefault(dayOfWeek,
-            DateTimeParser.parseIntegerToTime(DEFAULT_START_HOUR, LATENESS_MINUTE));
-        LocalTime absentTime = absentTimes.getOrDefault(dayOfWeek,
-            DateTimeParser.parseIntegerToTime(DEFAULT_START_HOUR, ABSENCE_MINUTE));
-
-        if (time.isAfter(absentTime)) {
-            return ABSENCE;
-        }
-        if (time.isAfter(lateTime)) {
-            return LATENESS;
-        }
-        return ATTENDANCE;
+        this.status = checkStatus(time, dayOfWeek);
     }
 
     public LocalTime getTime() {
