+++ src/main/java/domain/Penalty.java
@@ -1,6 +1,10 @@
 package domain;
 
+import java.time.LocalDate;
 import java.util.Arrays;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 public enum Penalty {
     EXPELLED("제적", 6),
@@ -22,5 +26,15 @@ public enum Penalty {
             .filter(penalty -> totalAbsences >= penalty.count)
             .findFirst()
             .orElse(NONE);
+    }
+
+    public static Map<String, StatisticsResult> calculateExpelledWarning
+        (LocalDate nowDate, Map<String, Crew> crews) {
+        return crews.entrySet().stream()
+            .map(entry -> Map.entry(entry.getKey(),
+                AttendanceStatus.countStatus(nowDate, entry.getValue())))
+            .filter(entry -> entry.getValue().hasPenalty())
+            .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue,
+                (a, b) -> b, LinkedHashMap::new));
     }
 }
